requirements.txt
python-pkcs11==0.7.0
pyHanko==0.21.0